SCB-2094 Move service/lock package
